refactor: useApproveCallback, useWrapCallback & useSwapCallback hook (#303)
* fix: restructure useApproveCallbackhook

* fix: delete useApproveCallback

* fix: restructure useWrapCallbackHook

* fix: wrap mesage change

* fix: resolve codesmell

* fix: restructure useSwapCallbackhook

* fix: delete useSwapCallbackhook file

* fix: commitlint issues
