+++ commitlint.config.js
@@ -11,20 +11,7 @@
 module.exports = {
   extends: ['@commitlint/config-conventional'],
   rules: {
-    'subject-case': [
-      2,
-      'always',
-      [
-        'lower-case', // default
-        'camel-case', // camelCase
-        'kebab-case', // kebab-case
-        'pascal-case', // PascalCase
-        'sentence-case', // Sentence case
-        'snake-case', // snake_case
-        'start-case', // Start Case,
-        'upper-case', // Upper case
-      ],
-    ],
+    'subject-case': [0, 'always'],
   },
   ignores: [(commit) => commit.includes('[skip ci]')],
 };
