+++ src/components/Pools/Stake/index.tsx
@@ -20,7 +20,8 @@ import { PNG } from 'src/constants/tokens';
 import { usePair } from 'src/data/Reserves';
 import { useChainId, useLibrary, usePangolinWeb3, useRefetchMinichefSubgraph } from 'src/hooks';
 import { MixPanelEvents, useMixpanel } from 'src/hooks/mixpanel';
-import { ApprovalState, useApproveCallback } from 'src/hooks/useApproveCallback';
+import { ApprovalState } from 'src/hooks/useApproveCallback/constant';
+import { useApproveCallback } from 'src/hooks/useApproveCallback/evm';
 import { usePairContract, useStakingContract } from 'src/hooks/useContract';
 import { useGetTransactionSignature } from 'src/hooks/useGetTransactionSignature';
 import useTransactionDeadline from 'src/hooks/useTransactionDeadline';
