+++ src/hooks/useApproveCallback/constant.ts
@@ -1,0 +1,6 @@
+export enum ApprovalState {
+  UNKNOWN,
+  NOT_APPROVED,
+  PENDING,
+  APPROVED,
+}
