+++ src/hooks/useSwapCallback/constant.ts
@@ -1,0 +1,5 @@
+export enum SwapCallbackState {
+  INVALID,
+  LOADING,
+  VALID,
+}
