+++ src/hooks/useSwapCallback/evm.ts
@@ -1,29 +1,18 @@
-/* eslint-disable max-lines */
 import { BigNumber } from '@ethersproject/bignumber';
 import { Contract } from '@ethersproject/contracts';
-import { parseUnits } from '@ethersproject/units';
-import { CAVAX, JSBI, Percent, Router, SwapParameters, Token, Trade, TradeType } from '@pangolindex/sdk';
+import { JSBI, Percent, Router, SwapParameters, Trade, TradeType } from '@pangolindex/sdk';
 import { useMemo } from 'react';
-import { NEAR_EXCHANGE_CONTRACT_ADDRESS } from 'src/connectors';
-import { INITIAL_ALLOWED_SLIPPAGE, ONE_YOCTO_NEAR, ZERO_ADDRESS } from 'src/constants';
+import { INITIAL_ALLOWED_SLIPPAGE, ZERO_ADDRESS } from 'src/constants';
 import { BIPS_BASE } from 'src/constants/swap';
-import { useGetNearPoolId } from 'src/data/Reserves';
+import { useDaasFeeTo } from 'src/state/pswap/hooks';
 import { useTransactionAdder } from 'src/state/ptransactions/hooks';
 import { calculateGasMargin, getRouterContract, getRouterContractDaaS, isAddress, shortenAddress } from 'src/utils';
-import { hederaFn } from 'src/utils/hedera';
 import isZero from 'src/utils/isZero';
-import { FunctionCallOptions, Transaction, nearFn } from 'src/utils/near';
-import { useDaasFeeTo } from '../state/pswap/hooks';
-import useENS from './useENS';
-import { Version } from './useToggledVersion';
-import useTransactionDeadline from './useTransactionDeadline';
-import { useChainId, useLibrary, usePangolinWeb3 } from './index';
-
-export enum SwapCallbackState {
-  INVALID,
-  LOADING,
-  VALID,
-}
+import { useChainId, useLibrary, usePangolinWeb3 } from '../index';
+import useENS from '../useENS';
+import { Version } from '../useToggledVersion';
+import useTransactionDeadline from '../useTransactionDeadline';
+import { SwapCallbackState } from './constant';
 
 interface SwapCall {
   contract: Contract;
@@ -245,215 +234,3 @@ export function useSwapCallback(
     };
   }, [trade, account, chainId, recipient, recipientAddressOrName, swapCalls, addTransaction]);
 }
-
-export function useNearSwapCallback(
-  trade: Trade | undefined, // trade to execute, required
-  recipientAddressOrName: string | null, // the ENS name or address of the recipient of the trade, or null if swap should be returned to sender
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _allowedSlippage: number = INITIAL_ALLOWED_SLIPPAGE, // in bips
-): { state: SwapCallbackState; callback: null | (() => Promise<string>); error: string | null } {
-  const { account } = usePangolinWeb3();
-  const chainId = useChainId();
-  const { library } = useLibrary();
-
-  const addTransaction = useTransactionAdder();
-
-  const poolId = useGetNearPoolId(trade?.inputAmount?.currency as Token, trade?.outputAmount?.currency as Token);
-
-  return useMemo(() => {
-    if (!trade || !library || !account || !chainId) {
-      return { state: SwapCallbackState.INVALID, callback: null, error: 'Missing dependencies' };
-    }
-
-    return {
-      state: SwapCallbackState.VALID,
-      callback: async function onSwap(): Promise<any> {
-        const transactions: Transaction[] = [];
-        const tokenInActions: FunctionCallOptions[] = [];
-
-        const inputToken = trade.inputAmount?.currency;
-        const outPutToken = trade.outputAmount?.currency;
-
-        const inputCurrencyId = inputToken instanceof Token ? inputToken?.address : undefined;
-        const outputCurrencyId = outPutToken instanceof Token ? outPutToken?.address : undefined;
-        const inputAmount = trade.inputAmount.toExact();
-
-        if (!inputCurrencyId || !outputCurrencyId) {
-          throw new Error(`Missing Currency`);
-        }
-
-        const tokenRegistered = await nearFn.getStorageBalance(outputCurrencyId).catch(() => {
-          throw new Error(`${outPutToken?.symbol} doesn't exist.`);
-        });
-
-        if (tokenRegistered === null) {
-          transactions.push({
-            receiverId: outputCurrencyId,
-            functionCalls: [
-              nearFn.storageDepositAction({
-                accountId: account,
-                registrationOnly: true,
-                amount: '0.00125',
-              }),
-            ],
-          });
-        }
-
-        const swapActions = {
-          pool_id: poolId,
-          token_in: inputCurrencyId,
-          token_out: outputCurrencyId,
-          amount_in: parseUnits(inputAmount, inputToken?.decimals).toString(),
-          min_amount_out: '0',
-        };
-
-        tokenInActions.push({
-          methodName: 'ft_transfer_call',
-          args: {
-            receiver_id: NEAR_EXCHANGE_CONTRACT_ADDRESS[chainId],
-            amount: parseUnits(inputAmount, inputToken?.decimals).toString(),
-            msg: JSON.stringify({
-              force: 0,
-              actions: [swapActions],
-            }),
-          },
-          amount: ONE_YOCTO_NEAR,
-        });
-
-        transactions.push({
-          receiverId: inputCurrencyId,
-          functionCalls: tokenInActions,
-        });
-
-        return nearFn.executeMultipleTransactions(transactions);
-      },
-      error: null,
-    };
-  }, [trade, poolId, library, account, chainId, recipientAddressOrName, addTransaction]);
-}
-
-// returns a function that will execute a swap, if the parameters are all valid
-// and the user has approved the slippage adjusted input amount for the trade
-export function useHederaSwapCallback(
-  trade: Trade | undefined, // trade to execute, required
-  recipientAddress: string | null, // the ENS name or address of the recipient of the trade, or null if swap should be returned to sender
-  allowedSlippage: number = INITIAL_ALLOWED_SLIPPAGE, // in bips
-): { state: SwapCallbackState; callback: null | (() => Promise<string>); error: string | null } {
-  const { account } = usePangolinWeb3();
-  const chainId = useChainId();
-
-  let deadline = useTransactionDeadline();
-
-  const currentTime = BigNumber.from(new Date().getTime());
-  if (deadline && deadline < currentTime.add(10)) {
-    deadline = currentTime.add(10);
-  }
-
-  const addTransaction = useTransactionAdder();
-
-  let recipient: string | null | undefined = '';
-
-  if (recipientAddress === null || !recipientAddress) {
-    recipient = account;
-  } else if (hederaFn.isHederaIdValid(recipientAddress)) {
-    recipient = hederaFn.idToAddress(recipientAddress);
-  }
-
-  return useMemo(() => {
-    if (!trade || !account || !chainId || !deadline) {
-      return { state: SwapCallbackState.INVALID, callback: null, error: 'Missing dependencies' };
-    }
-    if (!recipient) {
-      if (recipientAddress !== null) {
-        return { state: SwapCallbackState.INVALID, callback: null, error: 'Invalid recipient' };
-      } else {
-        return { state: SwapCallbackState.LOADING, callback: null, error: null };
-      }
-    }
-
-    const calls = Router.swapCallParameters(trade, {
-      feeOnTransfer: trade.tradeType === TradeType.EXACT_INPUT ? true : false,
-      allowedSlippage: new Percent(JSBI.BigInt(allowedSlippage), BIPS_BASE),
-      recipient,
-      deadline: deadline?.toNumber(),
-    });
-
-    const tradeVersion = Version.v2;
-
-    return {
-      state: SwapCallbackState.VALID,
-      callback: async function onSwap(): Promise<string> {
-        try {
-          if (!recipient) return '';
-
-          const inputCurrency = trade.inputAmount.currency;
-
-          const amountIn = trade.maximumAmountIn(new Percent(JSBI.BigInt(allowedSlippage), BIPS_BASE));
-          const amountOut = trade.minimumAmountOut(new Percent(JSBI.BigInt(allowedSlippage), BIPS_BASE));
-
-          const args = {
-            methodName: calls?.methodName,
-
-            HBARAmount: inputCurrency === CAVAX[chainId] ? amountIn?.toExact() : undefined,
-            tokenInAmount: amountIn?.raw?.toString(),
-            tokenOutAmount: amountOut?.raw?.toString(),
-            path: trade.route.path.map((token) => token.address), // token address
-            recipient,
-            account,
-            chainId,
-            deadline: deadline ? deadline?.toNumber() : 0,
-            exactAmountIn: trade.tradeType === TradeType.EXACT_INPUT,
-          };
-          const response = await hederaFn.swap(args);
-
-          if (response) {
-            const inputSymbol = trade.inputAmount.currency.symbol;
-            const outputSymbol = trade.outputAmount.currency.symbol;
-            const inputAmount = trade.inputAmount.toSignificant(3);
-            const outputAmount = trade.outputAmount.toSignificant(3);
-
-            const base = `Swap ${inputAmount} ${inputSymbol} for ${outputAmount} ${outputSymbol}`;
-
-            const withRecipient = recipient === account ? base : `${base} to ${recipientAddress}`;
-
-            const withVersion =
-              tradeVersion === Version.v2
-                ? withRecipient
-                : `${withRecipient} on ${(tradeVersion as any).toUpperCase()}`;
-
-            addTransaction(response, {
-              summary: withVersion,
-            });
-
-            return response.hash;
-          }
-
-          return '';
-        } catch (error: any) {
-          // if the user rejected the tx, pass this along
-          if (error?.code === 4001) {
-            throw new Error('Transaction rejected.');
-          } else {
-            // otherwise, the error was unexpected and we need to convey that
-            console.error(`Swap failed`, error);
-            throw new Error(`Swap failed: ${error.message}`);
-          }
-        }
-      },
-
-      error: null,
-    };
-  }, [trade, account, chainId, recipient, recipientAddress, addTransaction, deadline]);
-}
-
-export function useDummySwapCallback(
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _trade: Trade | undefined, // trade to execute, required
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _allowedSlippage: number = INITIAL_ALLOWED_SLIPPAGE, // in bips
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _recipientAddressOrName: string | null, // the ENS name or address of the recipient of the trade, or null if swap should be returned to sender
-): { state: SwapCallbackState; callback: null | (() => Promise<string>); error: string | null } {
-  return { state: SwapCallbackState.INVALID, callback: null, error: null };
-}
-/* eslint-enable max-lines */
