+++ src/hooks/useWrapCallback/constant.ts
@@ -1,0 +1,7 @@
+export enum WrapType {
+  NOT_APPLICABLE,
+  WRAP,
+  UNWRAP,
+}
+
+export const NOT_APPLICABLE = { wrapType: WrapType.NOT_APPLICABLE, executing: false };
