+++ src/state/psarstake/dummyHooks.ts
@@ -2,7 +2,7 @@ import { TokenAmount } from '@pangolindex/sdk';
 import { useState } from 'react';
 import { PNG } from 'src/constants/tokens';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
-import { ApprovalState } from 'src/hooks/useApproveCallback';
+import { ApprovalState } from 'src/hooks/useApproveCallback/constant';
 import { Position } from './types';
 
 /* eslint-disable @typescript-eslint/no-unused-vars */
